Refactor damage and movement methods for precision
Updated `takeDamage`, `move`, and `setPosition` methods in the `Enemy`, `entity`, and `player` classes to use `float` parameters for improved accuracy in damage calculations and movement. Added new methods in the `player` class for health management. Removed unnecessary debug output in `player::handleInput`. These changes enhance the overall robustness and flexibility of game mechanics.

+++ headers/Enemy.h
@@ -29,7 +29,7 @@ public:
 	Enemy(int id, float x, float y, int width, int height, int health, int damage, sf::Sprite sprite);
 	void attack(entity& target, float deltaTime);
 	void move(float deltaX, float deltaY);
-	void takeDamage(int damage);
+	void takeDamage(float damage);
 	void setHealth(int damage);
 	void die();
 	void setAggroRange(int range);
